package/weston: bump to version 14.0.1
From
https://lists.freedesktop.org/archives/wayland-devel/2024-October/043858.html:

This is Weston 14.0.1, a bug-fix release. Most of the fixes here are
addressing some build issues with a C++ compiler, but also addressing a
more critical issue that was causing the gtk4 test suite or other clients
to crash when using the headless backend.

Some distros already carry a patch to address that, but in case they
have not, this would be a good time to update.

David Edmundson (1):
      libweston: Send seat name before announcing devices

EatingSumo (1):
      libweston/screenshooter: Fix build when __builtin_clz is not available

Jan Alexander Steffens (heftig) (1):
      libweston/noop-renderer: Check shm_buffer for NULL

Jan Engelhardt (1):
      build: insert missing wayland-server-protocol dependency

Jeri Li (1):
      libweston/desktop: avoid weston crash while xdg_surface ack_configure

Marius Vlad (4):
      libweston: Move weston_get_backend_type after enum declaration
      desktop-shell: Don't attempt to re-add the view to panel layer
      libweston/matrix: Fix narrow conversion with C++
      build: bump to version 14.0.1 for the point release

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>