package/efl: bump to version 1.28.1
Some important addition and several bug fixes for gcc-15. Some of those
fixes are committed but not part of version 1.28.1 so they are added as
local patches while dropping old patches except the tslib one.

See Release Notes of all versions:
https://www.enlightenment.org/news/2025-03-17-efl-1.28.1
https://www.enlightenment.org/news/2025-01-11-efl-1.28.0
https://www.enlightenment.org/news/2023-12-23-efl-1.27.0

COPYING.images was removed as part of upstream commit
83d598a79c16b532113aeda8b1d3ff8a0816bde7.

Signed-off-by: Giulio Benetti <giulio.benetti@benettiengineering.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>